move selectStatusLoadingState to statuses selectors
